+++ src/coreComponents/constitutive/CMakeLists.txt
@@ -31,6 +31,7 @@ set( constitutive_headers
      fluid/MultiFluidExtrinsicData.hpp
      fluid/PhaseModel.hpp
      fluid/PVTDriver.hpp
+     fluid/PVTDriverRunTest.hpp
      fluid/PVTOData.hpp
      fluid/PVTFunctions/PhillipsBrineDensity.hpp
      fluid/PVTFunctions/PhillipsBrineViscosity.hpp
@@ -143,6 +144,9 @@ set( constitutive_sources
      fluid/DeadOilFluid.cpp
      fluid/MultiFluidBase.cpp
      fluid/PVTDriver.cpp
+     fluid/PVTDriverRunTestDeadOilFluid.cpp
+     fluid/PVTDriverRunTestCompositionalMultiphaseFluid.cpp
+     fluid/PVTDriverRunTestCO2Brine.cpp
      fluid/PVTOData.cpp
      fluid/PVTFunctions/PhillipsBrineDensity.cpp
      fluid/PVTFunctions/PhillipsBrineViscosity.cpp
